refactor(vcs): one push handler, parameterized by provider
The VCS refactor copied webhook-push.ts to vcs.webhook.processor.ts and
rewired the caller to the copy, leaving the original in the tree with zero
importers. The two differed only in imports, a threaded provider, and log
labels — but the orphan is the file carrying the cross-tenant installation
binding guard, delivery dedup and cloud forwarding, so any future fix to it
would have merged cleanly and had no effect, with no conflict to signal it.

handlePush now takes a provider and threads it through
BranchDeploymentTrigger: the delivery row's `source`, the compareCommits and
getRepository calls (via VcsStrategyFactory), and the actor-context labels.
The copy is deleted.

github.webhook.ts calls handlePush directly and drops the parseWebhookPayload
round-trip: it is a pure cast reached from inside `case "push":`, so its null
branch was unreachable.

No behavior change for GitHub — provider is "github" at the only call site.

bun run --cwd apps/api lint   → clean
bun run --cwd apps/api test   → 427 files, 5203 passed / 3 skipped
