+++ apps/api/src/modules/github/github.webhook.ts
@@ -18,8 +18,7 @@ import { verifyHmacSha256 } from "../webhooks/webhook.service";
 // resolveProjectWebhookSecret (github.service) was the old single-secret reader;
 // verify() now collects ALL candidate secrets via collectDeliverySecrets below.
 import { handleInstallation } from "./webhook-installation";
-import { VcsStrategyFactory } from "../vcs/vcs.factory";
-import { handlePush as handleVcsPush } from "../vcs/vcs.webhook.processor";
+import { handlePush } from "./webhook-push";
 import { handleCheckRun } from "./webhook-check-run";
 import type {
   WebhookProvider,
@@ -186,15 +185,7 @@ export const githubWebhookProvider: WebhookProvider = {
         result = await handleInstallation(payload as GitHubInstallationPayload);
         break;
       case "push":
-        const vcsPayload = VcsStrategyFactory.getStrategy("github").parseWebhookPayload(
-          payload,
-          "push",
-        );
-        if (!vcsPayload) {
-          result = { success: false, event: "push", message: "Failed to parse webhook payload" };
-          break;
-        }
-        result = await handleVcsPush("github", vcsPayload);
+        result = await handlePush("github", payload as GitHubPushPayload);
         break;
       case "check_run":
         result = await handleCheckRun(payload as GitHubCheckRunPayload);
